+++ exthost/action_network_tcp_reset.go
@@ -49,7 +49,7 @@ func getNetworkTcpResetDescription() action_kit_api.ActionDescription {
 				Type:        action_kit_api.ActionParameterTypeStringArray,
 				Required:    new(false),
 				Advanced:    new(true),
-				Order:       new(104),
+				Order:       new(106),
 			},
 		),
 	}